+++ projects/iad-interface-admin/partner/src/public_api.ts
@@ -1,15 +1,7 @@
 // #model
 export * from './lib/model/projection.model';
 
-// #service
-export * from './lib/services/data-chain.service';
-export * from './lib/services/document-info-buffer.service';
-export * from './lib/services/document-operations.service';
-export * from './lib/services/data-table-information.service';
-export * from './lib/services/actual-selection-chain.service';
-
 // #projection-table
-export * from './lib/projection-table/presentation-helper';
 export * from './lib/projection-table/projection-table.component';
 export * from './lib/projection-table/table-toolbar/table-toolbar.component';
 export * from './lib/projection-table/projection-table.module';